Test fix, presentation session rename
